wp-color-management-v1: allocate description ids linearly

+++ src/ifs/color_management/wp_color_management_surface_feedback_v1.rs
@@ -45,7 +45,7 @@ impl WpColorManagementSurfaceFeedbackV1 {
     pub fn send_preferred_changed(&self, cd: &ColorDescription) {
         self.client.event(PreferredChanged {
             self_id: self.id,
-            identity: cd.id.raw(),
+            identity: cd.id.raw() as u32,
         });
     }
 }
